alpaca_trading/requirements.txt
alpaca-trade-api
python-dotenv
tabulate